refactor: Extract geolocation, reverse geocoding, and city search logic into dedicated custom hooks.
